Add files via upload
This batch script is useful to transpose the contains of the file. 
For example, if the files contains the rows of data in an single column, this program will convert it into single row of multiple column and vice versa.
This is useful for giving variable directly to next script.