Add files via upload
The batch script rename_1000_file.sh is use for renaming the file in padded numerical format. That is file1.pdb, file10.pdb, and file1000.pdb will be renamed to file0001.pdb, file0010.pdb, and file1000.pdb respectively. 
In the place of <file_name>  add the name of the file and run the batch script.